+++ src/matrix/matrix_IO.cpp
@@ -57,7 +57,7 @@ void d_real_SpMat::input_mm(const char* filename){
      } while (buf[0] == '%');
 
      //check size
-     size_t rowN, colN, NNZ;
+     int rowN, colN, NNZ;
 
      std::istringstream data(buf);
      data >> rowN >> colN >> NNZ;
@@ -78,12 +78,12 @@ void d_real_SpMat::input_mm(const char* filename){
      row_ptr.resize(row+1, 0);
      col_ind.resize(nnz, 0);
 
-	 size_t c_row = 1;
+	 int c_row = 1;
 	 row_ptr[0] = 0;
 	 //input MM data as selected matrix storage format
- 	 for (size_t i = 0; i < NNZ; i++) {
-		 size_t idx, jdx;
-		 size_t value;
+ 	 for (int i = 0; i < NNZ; i++) {
+		 int idx, jdx;
+		 int value;
 
 		 getline(ifs, buf);
 		 std::istringstream data(buf);
@@ -115,8 +115,8 @@ void d_real_SpMat::output(){
          std::cerr << "This matrix is brank" << std::endl;
 	}
 
-	for(size_t i = 0; i < row; i++){
-		for(size_t j = row_ptr[i]; j < row_ptr[i+1]; j++){
+	for(int i = 0; i < row; i++){
+		for(int j = row_ptr[i]; j < row_ptr[i+1]; j++){
 			std::cout << i+1 << "\t" << col_ind[j]+1 << "\t" << val[j] << std::endl;
 		}
 	}
@@ -131,8 +131,8 @@ void d_real_SpMat::output_mm(const char* filename){
     out << (MM_BANNER " " MM_MAT " " MM_FMT " " MM_TYPE_REAL " " MM_TYPE_GENERAL) << std::endl;
     out << row << " " << row << " " << nnz << std::endl;
 
-	for(size_t i = 0; i < row; i++){
-		for(size_t j = row_ptr[i]; j < row_ptr[i+1]; j++){
+	for(int i = 0; i < row; i++){
+		for(int j = row_ptr[i]; j < row_ptr[i+1]; j++){
 			out << i+1 << " " << col_ind[j]+1 << " " << val[j] << std::endl;
 		}
 	}
