+++ src/vector/blas/axpy.cpp
@@ -12,7 +12,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -44,7 +44,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -75,7 +75,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -106,7 +106,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(0.0);
@@ -137,7 +137,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = broadcast(0.0);
@@ -169,7 +169,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = broadcast(0.0);
@@ -201,7 +201,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = broadcast(0.0);
@@ -230,7 +230,7 @@ namespace dd_avx{
 		}
 
 #pragma omp parallel for
-		for(size_t i = 0 ; i<y.size();i++){
+		for(int i = 0 ; i<y.size();i++){
 			y.data()[i] = y.data()[i] + alpha * x.data()[i];
 		}
 	}
